hands-on_code/semla_hands-on_coding-master/requirements.txt
tensorflow==1.12.0
matplotlib
pillow
scikit-image
scipy
